Move StorageVersion to restate-types and embed it in PartitionProcessorStatus
The previous commit exposed the partition-store storage version on the
status wire as a raw `Option<u32>`. That left two related warts:

  * `StorageVersion` lived in `restate-partition-store` and was reachable
    from `PartitionProcessorStatus` only via integer round-tripping, even
    though the type semantically belongs to the cluster vocabulary.
  * Restatectl / data-fusion consumers had no way to recover the typed
    enum from the wire value.

Move the enum into `restate-types::partitions::storage_version` and let
`PartitionProcessorStatus` hold it directly as `Option<StorageVersion>`.
The proto schema keeps `optional uint32 storage_version = 15`; a small
`storage_version_to_u32` map function feeds the prost conversion. On the
bilrost wire the new `bilrost::Enumeration` derive on `StorageVersion`
encodes the discriminant the same way an integer field would.

The migration methods (`run_migrations_up_to`, `do_migration`) stay in
`restate-partition-store`; only the data type relocates. `TryFrom<u16>`
now returns a small `UnknownStorageVersion` error that partition-store
wraps into `StorageError::Generic` at the boundary.

+++ crates/worker/src/partition/mod.rs
@@ -243,7 +243,7 @@ impl PartitionProcessorBuilder {
 
         let last_applied_log_lsn_watch = watch::Sender::new(Lsn::INVALID);
         // The storage version does not change after having opened the partition store
-        status.storage_version = Some(partition_store.storage_version() as u16);
+        status.storage_version = Some(partition_store.storage_version());
 
         Ok(PartitionProcessor {
             key_filter: KeyFilter::Within(partition_store.partition_key_range().into()),
